Added Elkarbackup v1.3.0 container. https://www.elkarbackup.org/ https://github.com/elkarbackup/elkarbackup
Elkarbackup is an open source backup solution for your network.
The image is the latest version which supports multiple backups locations.
It has some non important glitches, take a look to the docker readme at
https://github.com/elkarbackup/elkarbackup-docker/blob/master/elkarbackup/1.3/README.md

Two yml are included, one without proxy and one with an oauth2 proxy in front.
Also, the yml doesn't backups the elkarbackup db. It can be added on a similar way than nextcloud,
however elkarbackup provides a script that could be used to backup everything onto
another place.